eks/cdk/cdk-eks-cluster-stack/requirements.txt
aws-cdk-lib==2.179.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v33